[back-end] Moved http errors classes to httpErrors folder
