+++ backend/src/httpErrors/AuthenticationHttpError.ts
